% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_getranks.R
\name{.fp_string_to_json}
\alias{.fp_string_to_json}
\title{Extract JS from ECR}
\usage{
.fp_string_to_json(string)
}
\description{
Extract JS from ECR
}
\keyword{Converts}
\keyword{FantasyPros}
\keyword{JSON}
\keyword{internal}
\keyword{string}
\keyword{to}
