% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{ffpros-package}
\alias{ffpros}
\alias{ffpros-package}
\title{ffpros: Retrieves Data from FantasyPros}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Helps gather data from FantasyPros.com with
    rate-limiting/caching, forming appropriate calls, and returning tidy
    dataframes which can be easily connected to other data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ffpros.dynastyprocess.com}
  \item \url{https://github.com/dynastyprocess/ffpros}
  \item Report bugs at \url{https://github.com/dynastyprocess/ffpros/issues}
}

}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca}

}
\keyword{internal}
