% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_projections.R
\name{fp_projections}
\alias{fp_projections}
\title{Import FantasyPros Projections}
\usage{
fp_projections(page, ..., sport = NULL, include_metadata = NULL)
}
\arguments{
\item{page}{which page}

\item{...}{named arguments to pass as query parameters}

\item{sport}{the sport to look up, defaults to nfl}

\item{include_metadata}{logical - default FALSE (only return tibble), if TRUE returns list with both tibble and metadata}
}
\description{
Scrapes FantasyPros ranks from the specified page (and optionally, returns metadata).
}
\examples{
\donttest{
if(interactive()){
fp_projections("qb")
fp_projections("wr", year = 2016, week = 2, scoring = "PPR")
fp_projections("flex", year = 2020, week = 7, scoring = "PPR", `min-yes`="true", `max-yes`="true")
}
}

}
\seealso{
\url{https://github.com/dynastyprocess/data}

\url{https://www.fantasypros.com}
}
