% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_rankings_generic.R
\name{fp_rankings}
\alias{fp_rankings}
\title{Import latest FantasyPros Rankings}
\usage{
fp_rankings(page, ..., sport = NULL, include_metadata = NULL)
}
\arguments{
\item{page}{which page}

\item{...}{named arguments to pass as query parameters}

\item{sport}{the sport to look up, defaults to nfl}

\item{include_metadata}{logical - default FALSE (only return tibble), if TRUE returns list with both tibble and metadata}
}
\description{
Scrapes FantasyPros ranks from the specified page (and optionally, returns metadata).
}
\examples{
\donttest{
fp_rankings(page = "dynasty-overall", sport = "nfl")
}

}
\seealso{
\url{https://github.com/dynastyprocess/data}

\url{https://www.fantasypros.com}
}
