% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_config.R
\name{fp_set_metadata}
\alias{fp_set_metadata}
\title{Set whether metadata is returned}
\usage{
fp_set_metadata(include = FALSE)
}
\arguments{
\item{include}{logical}
}
\description{
This option allows you to configure whether the output of functions includes metadata,
or is simplified to only the dataframe. Defaults to simplifying (i.e. include = FALSE)
}
