% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{fp_set_ratelimit}
\alias{fp_set_ratelimit}
\title{Set rate limit}
\usage{
fp_set_ratelimit(rate_number = 1, rate_seconds = 5, rate_limit = TRUE)
}
\arguments{
\item{rate_number}{number of calls per \code{rate_seconds}}

\item{rate_seconds}{number of seconds}

\item{rate_limit}{a logical to turn rate limiting on if TRUE and off if FALSE}
}
\description{
A helper function that adjusts rate limits for ffpros.
}
