% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{fp_set_sport}
\alias{fp_set_sport}
\title{Set default sport}
\usage{
fp_set_sport(sport)
}
\arguments{
\item{sport}{one of c("nfl","nba","nhl","mlb")}
}
\description{
A helper function that sets the default sport for the ffpros functions.
Can be overridden directly.
}
