% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{fp_set_useragent}
\alias{fp_set_useragent}
\title{Set user agent}
\usage{
fp_set_useragent(user_agent)
}
\arguments{
\item{user_agent}{A character string to use as the user agent}
}
\description{
Configures the user agent for the ffpros scraping function.
}
