% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.fp_string_to_json}
\alias{.fp_string_to_json}
\title{Converts FantasyPros string to JSON}
\usage{
.fp_string_to_json(string, extractor = "ecrData")
}
\description{
Converts FantasyPros string to JSON
}
\keyword{internal}
