% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_getranks.R
\name{fp_rankings}
\alias{fp_rankings}
\title{Import latest FantasyPros Rankings}
\usage{
fp_rankings(page, ..., sport = "nfl", .return = "table")
}
\arguments{
\item{page}{which page}

\item{...}{named arguments to pass as query parameters}

\item{sport}{which sport, defaults to NFL}

\item{.return}{one of c("table","metadata","all"), defaults to table}
}
\value{
rankings data from FantasyPros
}
\description{
Scrapes FantasyPros ranks from the specified page and returns metadata
}
\examples{
\donttest{
fp_rankings(page = "dynasty-overall", sport = "nfl", return = "table")
}

}
\seealso{
\url{https://github.com/DynastyProcess/data/}

\url{https://www.fantasypros.com}
}
