% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_projections.R, R/fp_rankings.R
\name{fp_projections_parse.fp_nfl}
\alias{fp_projections_parse.fp_nfl}
\alias{fp_rankings_parse}
\alias{fp_rankings_parse.fp_nfl}
\alias{fp_rankings_parse.fp_nhl}
\alias{fp_rankings_parse.fp_nba}
\alias{fp_rankings_parse.fp_mlb}
\title{NFL method for fp_rankings}
\usage{
\method{fp_projections_parse}{fp_nfl}(response)

fp_rankings_parse(response)

\method{fp_rankings_parse}{fp_nfl}(response)

\method{fp_rankings_parse}{fp_nhl}(response)

\method{fp_rankings_parse}{fp_nba}(response)

\method{fp_rankings_parse}{fp_mlb}(response)
}
\description{
NFL method for fp_rankings

Dispatch fp_rankings_parse

NFL method for fp_rankings

NHL method for fp_rankings

NBA method for fp_rankings

MLB method for fp_rankings
}
\section{Methods (by class)}{
\itemize{
\item \code{fp_nfl}: 

\item \code{fp_nfl}: 

\item \code{fp_nhl}: 

\item \code{fp_nba}: 

\item \code{fp_mlb}: 
}}

\seealso{
\url{https://www.fantasypros.com/nhl/rankings/overall.php}

\url{https://www.fantasypros.com/nba/rankings/ros-overall.php}

\url{https://www.fantasypros.com/nba/rankings/ros-overall.php}
}
\keyword{internal}
